refactor(habits): separate concerns into modular architecture
- Split storage.ts into focused modules:
  - habitStorage.ts: Core habit CRUD operations
  - habitTemplates.ts: Template system with presets
  - storage.ts: Core storage utilities only

- Update imports across components and tests
- Maintain 98 passing tests with proper module isolation
- Follow separation of concerns principle for maintainability

This improves code organization and makes the codebase more
modular, extensible, and easier to maintain.
